convert setu model to typescript
